organized project by creating sub-directories
