+++ app/app.py
@@ -1,6 +1,7 @@
 from flask import Flask
-from app.db import mysql_uri
-from app.schema import *
+
+from app.database.db import mysql_uri
+from app.database.schema import *
 
 """
 This file is for running the app. To run the app, type "flask run" in the terminal
@@ -9,13 +10,13 @@ This file is for running the app. To run the app, type "flask run" in the termin
 app = Flask(__name__)
 
 # register Blueprints
-from app.vending_machine import vending_machine
+from app.endpoints.vending_machine import vending_machine
 
 app.register_blueprint(vending_machine)
-from app.products import products
+from app.endpoints.products import products
 
 app.register_blueprint(products)
-from app.machine_stocks import machine_stocks
+from app.endpoints.machine_stocks import machine_stocks
 
 app.register_blueprint(machine_stocks)
