+++ app/endpoints/machine_stocks.py
@@ -1,5 +1,6 @@
-from app.queryUtils import *
-from flask import Blueprint, request, jsonify, redirect, url_for, flash
+from flask import Blueprint, request, jsonify, redirect, url_for
+
+from app.database.queryUtils import *
 
 '''
 this file contains all function regarding CRUD operation for machine_stock table
@@ -13,6 +14,7 @@ query_strings - the query strings which are passed in as argument in the url
 return true if all criteria are passed else return false
 '''
 
+
 def validate_product_and_machine(query_strings):
     query_strings_are_valid = are_all_query_string_present(query_strings, ("machine_id", "product_id", "quantity"))
     no_duplicate_product_in_the_same_machine = not is_exist(MachineStock, {"product_id": query_strings["product_id"],
@@ -20,6 +22,7 @@ def validate_product_and_machine(query_strings):
     product_exists = is_exist(Products, {"id": query_strings["product_id"]})
     machine_exists = is_exist(vendingMachine, {"id": query_strings["machine_id"]})
     return product_exists and machine_exists and no_duplicate_product_in_the_same_machine and query_strings_are_valid
+
 
 def validate_product_quantity(query_strings, product_available):
     quantity_not_negative = int(query_strings["quantity"]) >= 0
@@ -29,9 +32,10 @@ def validate_product_quantity(query_strings, product_available):
         product_is_enough = int(product.product_quantity) >= int(query_strings["quantity"])
     return product_is_enough and quantity_not_negative
 
+
 def add_validate(query_strings):
     product_and_machine_available = validate_product_and_machine(query_strings)
-    product_quantity_is_valid = validate_product_quantity(query_strings,product_and_machine_available)
+    product_quantity_is_valid = validate_product_quantity(query_strings, product_and_machine_available)
     return product_and_machine_available and product_quantity_is_valid
